only serve graph in cgo build
